+++ examples/template-hydrogen-default/src/App.client.jsx
@@ -1,8 +1,0 @@
-import CartProvider from './components/CartProvider.client';
-
-/**
- *  Setup client context, though the children are most likely server components
- */
-export default function ClientApp({children}) {
-  return <CartProvider>{children}</CartProvider>;
-}