+++ packages/cli/src/commands/create/app/templates/index-html.ts
@@ -13,7 +13,7 @@ export default function ({ifFeature}: TemplateOptions) {
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/entry-client.jsx"></script>
+    <script type="module" src="/@shopify/hydrogen/entry-client"></script>
   </body>
 </html>
 `;