+++ packages/hydrogen/src/components/SelectedVariantImage/examples/selected-variant-image.example.tsx
@@ -1,9 +1,0 @@
-import {SelectedVariantImage, ProductProvider} from '@shopify/hydrogen';
-
-export function ProductDetails({product}) {
-  return (
-    <ProductProvider data={product}>
-      <SelectedVariantImage />
-    </ProductProvider>
-  );
-}