+++ packages/hydrogen/src/foundation/Helmet/index.ts
@@ -1,1 +1,0 @@
-export {Helmet, RealHelmetData} from './Helmet.client';