+++ packages/hydrogen/src/framework/Hydration/ServerComponentRequest.server.ts
@@ -1,7 +1,6 @@
 import type {ShopifyContextValue} from '../../foundation/ShopifyProvider/types';
 import {getTime} from '../../utilities/timing';
-import {HelmetData} from 'react-helmet-async';
-import type {RealHelmetData} from '../../foundation/Helmet';
+import {HelmetData as HeadData} from 'react-helmet-async';
 import {QueryCacheControlHeaders} from '../../utilities/log/log-cache-header';
 import {QueryTiming} from '../../utilities/log/log-query-timeline';
 
@@ -28,7 +27,7 @@ export class ServerComponentRequest extends Request {
   // CFW Request has a reserved 'context' property, use 'ctx' instead.
   public ctx: {
     cache: Map<string, any>;
-    helmet: RealHelmetData;
+    head: HeadData;
     shopifyConfig?: ShopifyContextValue;
     queryCacheControl: Array<QueryCacheControlHeaders>;
     queryTimings: Array<QueryTiming>;
@@ -56,7 +55,7 @@ export class ServerComponentRequest extends Request {
 
     this.ctx = {
       cache: new Map(),
-      helmet: new HelmetData({}) as unknown as RealHelmetData,
+      head: new HeadData({}),
       queryCacheControl: [],
       queryTimings: [],
     };