+++ packages/hydrogen/src/framework/plugin.ts
@@ -1,12 +1,16 @@
 import type {HydrogenVitePluginOptions, ShopifyConfig} from '../types';
 import hydrogenConfig from './plugins/vite-plugin-hydrogen-config';
 import type {Plugin} from 'vite';
-import hydrogenMiddleware from './plugins/vite-plugin-hydrogen-middleware';
+import hydrogenMiddleware, {
+  HYDROGEN_DEFAULT_SERVER_ENTRY,
+} from './plugins/vite-plugin-hydrogen-middleware';
 import hydrogenClientMiddleware from './plugins/vite-plugin-hydrogen-client-middleware';
+import platformEntry from './plugins/vite-plugin-platform-entry';
 // @ts-ignore
 import rsc from '@shopify/hydrogen/vendor/react-server-dom-vite/plugin';
 import ssrInterop from './plugins/vite-plugin-ssr-interop';
 import purgeQueryCache from './plugins/vite-plugin-purge-query-cache';
+import hydrationAutoImport from './plugins/vite-plugin-hydration-auto-import';
 import inspect from 'vite-plugin-inspect';
 import react from '@vitejs/plugin-react';
 import path from 'path';
@@ -22,6 +26,7 @@ export default (
     hydrogenClientMiddleware(),
     hydrogenMiddleware(shopifyConfig, pluginOptions),
     react(),
+    hydrationAutoImport(),
     ssrInterop(),
     rsc({
       clientComponentPaths: [
@@ -29,11 +34,20 @@ export default (
           path.dirname(require.resolve('@shopify/hydrogen/package.json'))
         ),
       ],
-      isServerComponentImporterAllowed(importer: string) {
-        return /(handle-worker-event|index)\.js/.test(importer);
+      isServerComponentImporterAllowed(importer: string, source: string) {
+        // Always allow the entry server (e.g. App.server.jsx) to be imported
+        // in other files such as worker.js or server.js.
+        const entryServer =
+          process.env.HYDROGEN_SERVER_ENTRY || HYDROGEN_DEFAULT_SERVER_ENTRY;
+
+        return (
+          source.includes(entryServer) ||
+          // TODO update this after handleEvent is replaced with handleRequest
+          /(handle-worker-event|index|entry-server)\.js/.test(importer)
+        );
       },
     }),
-
+    platformEntry(),
     pluginOptions.purgeQueryCacheOnBuild && purgeQueryCache(),
   ] as Plugin[];
 };