+++ packages/hydrogen/web-polyfills.js
@@ -1,0 +1,1 @@
+require('./dist/node/utilities/web-api-polyfill');