+++ packages/playground/server-components/worker.js
@@ -1,13 +1,16 @@
-import handleEvent from '@shopify/hydrogen/worker';
-import entrypoint from './src/entry-server.jsx';
-// eslint-disable-next-line node/no-missing-import
-import indexHtml from './dist/client/index.html?raw';
 import {getAssetFromKV} from '@cloudflare/kv-asset-handler';
+import handleRequest from './src/App.server';
+// eslint-disable-next-line node/no-missing-import
+import indexTemplate from './dist/client/index.html?raw';
 
 // Mock Oxygen global
 globalThis.Oxygen = {env: globalThis};
 
-async function assetHandler(event, url) {
+function isAsset(url) {
+  return /\.(png|jpe?g|gif|css|js|svg|ico|map)$/i.test(url.pathname);
+}
+
+async function handleAsset(url, event) {
   const response = await getAssetFromKV(event, {});
 
   if (response.status < 400) {
@@ -16,7 +19,7 @@ async function assetHandler(event, url) {
     const maxAge =
       filename.split('.').length > 2
         ? 31536000 // hashed asset, will never be updated
-        : 86400; // favico and other public assets
+        : 86400; // favicon and other public assets
 
     response.headers.append('cache-control', `public, max-age=${maxAge}`);
   }
@@ -24,24 +27,21 @@ async function assetHandler(event, url) {
   return response;
 }
 
-addEventListener('fetch', (event) => {
+async function handleEvent(event) {
   try {
-    event.respondWith(
-      handleEvent(event, {
-        entrypoint,
-        indexTemplate: indexHtml,
-        assetHandler,
-        cache: caches.default,
-        context: {
-          waitUntil: event.waitUntil ? (p) => event.waitUntil(p) : undefined,
-        },
-      })
-    );
+    const url = new URL(event.request.url);
+
+    if (isAsset(url)) {
+      return await handleAsset(url, event);
+    }
+
+    return await handleRequest(event.request, {
+      indexTemplate,
+      cache: caches.default,
+      context: event,
+    });
   } catch (error) {
-    event.respondWith(
-      new Response(error.message || error.toString(), {
-        status: 500,
-      })
-    );
+    return new Response(error.message || error.toString(), {status: 500});
   }
-});
+}
+addEventListener('fetch', (event) => event.respondWith(handleEvent(event)));