+++ scripts/jest-e2e-global-setup.ts
@@ -7,7 +7,9 @@ const DIR = join(tmpdir(), 'jest_playwright_global_setup');
 
 export default async function () {
   const browserServer = await chromium.launchServer({
+    // Change to `true` to see the chromium browser during tests
     headless: !process.env.VITE_DEBUG_SERVE,
+
     args: process.env.CI
       ? ['--no-sandbox', '--disable-setuid-sandbox']
       : undefined,